Split test_permission_group_delete.py test file
